Try the cargo workspace
