+++ .gitignore
@@ -1,2 +1,5 @@
 /target
 /test
+/src_old
+v0.3_notes.md
+README_v0.3.md
