+++ Cargo.toml
@@ -1,22 +1,7 @@
-[package]
-name = "pycrucible"
-version = "0.2.9"
-edition = "2024"
-
-[dependencies]
-walkdir = "2.4"
-maplit = "1.0"
-lazy_static = "1.4"
-once_cell = "1.18"
-toml = "0.8"
-clap = { version = "4.2", features = ["derive"] }
-serde = { version = "1.0", features = ["derive"] }
-reqwest = { version = "0.12", features = ["blocking", "rustls-tls"], default-features = false }
-spinners = "4"
-zip = "3"
-tar = "0.4"
-glob = "0.3"
-tempfile = "3"
-flate2 = "1"
-git2 = { version = "0.20", default-features = false }
-which = "8.0.0"
+[workspace]
+resolver = "3"
+members = [
+    "shared",
+    "pycrucible",
+    "runner"
+]
