+++ shared/src/spinner_utils.rs
